requirements.txt
anyio==4.2.0
asgiref==3.8.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
mysqlclient==2.2.4
packaging==23.2
pillow==10.2.0
pytchat==0.5.5
pytesseract==0.3.10
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
